# frontend/packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NookOS API",
    "description": "The open workspace operating system for modern software teams.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/dev-login": {
      "post": {
        "tags": [
          "crate::routes::auth"
        ],
        "summary": "POST /api/v1/auth/dev-login — dev/CI escape hatch. Compiled in, but\nhard-refused unless AUTH_DEV_MODE=true (and never in production).",
        "operationId": "dev_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "403": {
            "description": "dev mode disabled"
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "crate::routes::auth"
        ],
        "summary": "GET /api/v1/auth/me",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "not signed in"
          }
        }
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "tags": [
          "crate::routes::sessions"
        ],
        "operationId": "list_sessions",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/WorkspaceId"
                }
              ]
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Only sessions that are starting/running/detached.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::sessions"
        ],
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "workspace_id",
          "node_id",
          "runtime"
        ],
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "node_id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pin the session to a specific checkout path (e.g. a worktree). When\nomitted, the workspace's first checkout on the node is used."
          },
          "runtime": {
            "type": "string"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/WorkspaceId"
          }
        }
      },
...
      "DevLoginRequest": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "description": "The signed-in caller with their tenant.",
        "required": [
          "user",
          "tenant"
        ],
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/Tenant"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "Session": {
        "type": "object",
        "description": "Status values: `starting` | `running` | `detached` | `exited` | `error`.\nRuntime is an open string: \"claude\", \"hermes\", \"codex\", \"bash\", \"zsh\", ...",
        "required": [
          "id",
          "tenant_id",
          "workspace_id",
          "node_id",
          "name",
          "runtime",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "ended_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/SessionId"
          },
          "name": {
            "type": "string"
          },
          "node_id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "runtime": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "tmux_session": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/WorkspaceId"
          }
        }
      },
      "SessionId": {
        "type": "string",
        "format": "uuid"
      },
...
      "Tenant": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TenantId": {
        "type": "string",
        "format": "uuid"
      },
...
      "User": {
        "type": "object",
        "description": "Role values: `owner` | `admin` | `member` (TEXT CHECK in the schema).",
        "required": [
          "id",
          "tenant_id",
          "display_name",
          "email",
          "role",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/UserId"
          },
          "role": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserId": {
        "type": "string",
        "format": "uuid"
      },
...
      "WorkspaceId": {
        "type": "string",
        "format": "uuid"
      },
...
    }
  }
}